C functions

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 int count = 0;
-unsigned long long int a = 0, b = 1, c = 0;
+unsigned int a = 0, b = 1, c = 0;
 
 for (count = 0; count <= 98; count++)
 {
